+++ packages/ice/src/constant.ts
@@ -5,6 +5,8 @@ export const ROUTER_MANIFEST = path.join(RUNTIME_TMP_DIR, 'route-manifest.json')
 export const SERVER_ENTRY = path.join(RUNTIME_TMP_DIR, 'entry.server.ts');
 export const DATA_LOADER_ENTRY = path.join(RUNTIME_TMP_DIR, 'data-loader.ts');
 export const SERVER_OUTPUT_DIR = 'server';
+export const IMPORT_META_TARGET = 'import.meta.target';
+export const IMPORT_META_RENDERER = 'import.meta.renderer';
 export const CACHE_DIR = path.join('node_modules', '.cache');
 export const BUILDIN_ESM_DEPS = [
   '@ice/runtime',
@@ -12,8 +14,6 @@ export const BUILDIN_ESM_DEPS = [
 export const BUILDIN_CJS_DEPS = [
   'react',
   'react-dom',
-  '@uni/env',
-  'universal-env',
 ];
 
 export const WEB = 'web';
@@ -24,12 +24,13 @@ export const BYTEDANCE_MICROAPP = 'bytedance-microapp';
 export const BAIDU_SMARTPROGRAM = 'baidu-smartprogram';
 export const KUAISHOU_MINIPROGRAM = 'kuaishou-miniprogram';
 
-export const MINIAPP_PLATFORMS = [
+export const MINIAPP_TARGETS = [
   ALI_MINIAPP, WECHAT_MINIPROGRAM, BYTEDANCE_MICROAPP,
   BAIDU_SMARTPROGRAM, KUAISHOU_MINIPROGRAM,
 ];
-export const ALL_PLATFORMS = [
+
+export const TARGETS = [
   WEB,
   WEEX,
-  ...MINIAPP_PLATFORMS,
+  ...MINIAPP_TARGETS,
 ];