+++ packages/ice/src/service/serverCompiler.ts
@@ -22,6 +22,7 @@ import getServerEntry from '../utils/getServerEntry.js';
 import type { DepScanData } from '../esbuild/scan.js';
 import formatPath from '../utils/formatPath.js';
 import { createLogger } from '../utils/logger.js';
+import { getExpandedEnvs } from '../utils/runtimeEnv.js';
 import { scanImports } from './analyze.js';
 import type { DepsMetaData } from './preBundleCJSDeps.js';
 import preBundleCJSDeps from './preBundleCJSDeps.js';
@@ -71,18 +72,29 @@ export function createServerCompiler(options: Options) {
     compilationInfo,
     redirectImports,
     removeOutputs,
+    runtimeDefineVars = {},
     enableEnv = false,
     transformEnv = true,
   } = {}) => {
     let depsMetadata: DepsMetaData;
     let swcOptions = merge({}, {
       // Only get the `compilationConfig` from task config.
-      compilationConfig: {
-        ...(task.config?.swcOptions?.compilationConfig || {}),
-        // Force inline when use swc as a transformer.
-        sourceMaps: sourceMap && 'inline',
-      },
+      compilationConfig: getCompilationConfig(),
     }, swc);
+    function getCompilationConfig() {
+      const customCompilationConfig = task.config?.swcOptions?.compilationConfig || {};
+      const getConfig = typeof customCompilationConfig === 'function'
+        ? customCompilationConfig
+        : () => customCompilationConfig;
+
+      return (source, id) => {
+        return {
+          ...getConfig(source, id),
+          // Force inline when use swc as a transformer.
+          sourceMaps: sourceMap && 'inline',
+        };
+      };
+    }
     const enableSyntaxFeatures = syntaxFeatures && Object.keys(syntaxFeatures).some(key => syntaxFeatures[key]);
     const transformPlugins = getCompilerPlugins({
       ...task.config,
@@ -109,7 +121,6 @@ export function createServerCompiler(options: Options) {
     }
 
     // get runtime variable for server build
-    const runtimeDefineVars = {};
     Object.keys(process.env).forEach((key) => {
       // Do not transform env when bundle client side code.
       if (/^ICE_CORE_/i.test(key) && transformEnv) {
@@ -123,6 +134,18 @@ export function createServerCompiler(options: Options) {
       ...defineVars,
       ...runtimeDefineVars,
     };
+    const expandedEnvs = getExpandedEnvs();
+    // Add user defined envs.
+    for (const [key, value] of Object.entries(expandedEnvs)) {
+      define[`import.meta.env.${key}`] = JSON.stringify(value);
+    }
+    // Add process.env.
+    Object.keys(process.env)
+      .filter((key) => /^ICE_/.test(key) || key === 'NODE_ENV')
+      .forEach((key) => {
+        define[`import.meta.env.${key}`] = JSON.stringify(process.env[key]);
+      });
+
     const format = customBuildOptions?.format || 'esm';
 
     let buildOptions: esbuild.BuildOptions = {
@@ -141,6 +164,12 @@ export function createServerCompiler(options: Options) {
       define,
       absWorkingDir: rootDir,
       external: Object.keys(externals),
+      banner: customBuildOptions.platform === 'node' && server?.format !== 'cjs'
+        ? {
+            // See https://github.com/evanw/esbuild/issues/1921#issuecomment-1152991694
+            js: 'import { createRequire } from \'module\';const require = createRequire(import.meta.url);',
+          }
+        : undefined,
       plugins: [
         ...(customBuildOptions.plugins || []),
         emptyCSSPlugin(),