+++ packages/ice/src/utils/getServerCompilerPlugin.ts
@@ -15,9 +15,21 @@ interface Options {
   dataCache: Map<string, string>;
   serverCompileTask: ExtendsPluginAPI['serverCompileTask'];
   ensureRoutesConfig: () => Promise<void>;
+  runtimeDefineVars: Record<string, string>;
 }
+
 function getServerCompilerPlugin(serverCompiler: ServerCompiler, options: Options) {
-  const { outputDir, rootDir, serverEntry, userConfig, dataCache, serverCompileTask, ensureRoutesConfig, incremental } = options;
+  const {
+    outputDir,
+    rootDir,
+    serverEntry,
+    userConfig,
+    dataCache,
+    serverCompileTask,
+    ensureRoutesConfig,
+    runtimeDefineVars,
+    incremental,
+  } = options;
   const entryPoint = getServerEntry(rootDir, serverEntry);
   const { ssg, ssr, server: { format } } = userConfig;
   const isEsm = userConfig?.server?.format === 'esm';
@@ -29,10 +41,7 @@ function getServerCompilerPlugin(serverCompiler: ServerCompiler, options: Option
         outdir: path.join(outputDir, SERVER_OUTPUT_DIR),
         splitting: isEsm,
         format,
-        // When format set to esm and platform set to node,
-        // will cause error `Dynamic require of "XXX" is not supported`.
-        // https://github.com/evanw/esbuild/issues/1921
-        platform: isEsm ? 'browser' : 'node',
+        platform: 'node',
         mainFields: ['module', 'main'],
         outExtension: { '.js': isEsm ? '.mjs' : '.cjs' },
         metafile: true,
@@ -43,12 +52,12 @@ function getServerCompilerPlugin(serverCompiler: ServerCompiler, options: Option
         preBundle: format === 'esm' && (ssr || ssg),
         swc: {
           keepExports: (!ssg && !ssr) ? ['pageConfig'] : null,
-          keepPlatform: 'node',
           getRoutePaths: () => {
             return getRoutePathsFromCache(dataCache);
           },
         },
         removeOutputs: true,
+        runtimeDefineVars,
       },
     ],
     ensureRoutesConfig,