+++ packages/webpack-config/src/index.ts
@@ -1,7 +1,6 @@
 import * as path from 'path';
 import { createRequire } from 'module';
 import fg from 'fast-glob';
-// FIXME when prepack @pmmmwh/react-refresh-webpack-plugin
 import ReactRefreshWebpackPlugin from '@ice/bundles/compiled/@pmmmwh/react-refresh-webpack-plugin/lib/index.js';
 import bundleAnalyzer from '@ice/bundles/compiled/webpack-bundle-analyzer/index.js';
 import lodash from '@ice/bundles/compiled/lodash/index.js';
@@ -16,6 +15,7 @@ import type { Config, ModifyWebpackConfig } from './types.js';
 import configAssets from './config/assets.js';
 import configCss from './config/css.js';
 import AssetsManifestPlugin from './webpackPlugins/AssetsManifestPlugin.js';
+import EnvReplacementPlugin from './webpackPlugins/EnvReplacementPlugin.js';
 import getCompilerPlugins from './getCompilerPlugins.js';
 import getSplitChunksConfig, { FRAMEWORK_BUNDLES } from './config/splitChunks.js';
 import compilationPlugin from './unPlugins/compilation.js';
@@ -34,8 +34,10 @@ interface GetWebpackConfigOptions {
   webpack: typeof webpack;
   runtimeTmpDir: string;
   userConfigHash: string;
+  getExpandedEnvs: () => Record<string, string>;
+  runtimeDefineVars?: Record<string, any>;
 }
-type GetWebpackConfig = (options: GetWebpackConfigOptions) => Configuration;
+
 enum JSMinifier {
   terser = 'terser',
   swc = 'swc',
@@ -60,10 +62,72 @@ function getEntry(rootDir: string, runtimeTmpDir: string) {
   };
 }
 
-const getWebpackConfig: GetWebpackConfig = ({ rootDir, config, webpack, runtimeTmpDir, userConfigHash }) => {
+// format alias
+function getAliasWithRoot(rootDir: string, alias?: Record<string, string | boolean>) {
+  const aliasWithRoot = {};
+  Object.keys(alias).forEach((key) => {
+    const aliasValue = alias[key];
+    aliasWithRoot[key] = (aliasValue && typeof aliasValue === 'string' && aliasValue.startsWith('.')) ? path.join(rootDir, aliasValue) : aliasValue;
+  });
+  return aliasWithRoot;
+}
+
+const RUNTIME_PREFIX = /^ICE_/i;
+
+function getDefineVars(
+  config: Config,
+  runtimeDefineVars: Record<string, any>,
+  getExpandedEnvs: () => Record<string, string>,
+  webpack,
+) {
+  const { define = {} } = config;
+  // auto stringify define value
+  const defineVars = {};
+  Object.keys(define).forEach((key) => {
+    defineVars[key] = JSON.stringify(define[key]);
+  });
+
+  Object.keys(process.env).filter((key) => {
+    return RUNTIME_PREFIX.test(key) || ['NODE_ENV'].includes(key);
+  }).forEach((key) => {
+    runtimeDefineVars[`process.env.${key}`] =
+      /^ICE_CORE_/i.test(key)
+        // ICE_CORE_* will be updated dynamically, so we need to make it effectively
+        ? webpack.DefinePlugin.runtimeValue(() => JSON.stringify(process.env[key]), true)
+        : JSON.stringify(process.env[key]);
+  });
+  // ImportMeta.env is ice defined env variables.
+  runtimeDefineVars['import.meta.env'] = getImportMetaEnv(getExpandedEnvs);
+
+  return {
+    ...define,
+    ...runtimeDefineVars,
+  };
+}
+
+function getImportMetaEnv(getExpandedEnvs: () => Record<string, string>): Record<string, string> {
+  const env = {};
+  const validEnvKeys = ['NODE_ENV'];
+
+  Object.keys(process.env)
+    .filter((key) => RUNTIME_PREFIX.test(key) || validEnvKeys.includes(key))
+    .forEach((key) => {
+      env[key] = JSON.stringify(process.env[key]);
+    });
+
+  // User defined envs at `.env` series files.
+  const expandedEnvs = getExpandedEnvs();
+  for (const [key, value] of Object.entries(expandedEnvs)) {
+    env[`import.meta.env.${key}`] = JSON.stringify(value);
+  }
+
+  return env;
+}
+
+export function getWebpackConfig(options: GetWebpackConfigOptions): Configuration {
+  const { rootDir, config, webpack, runtimeTmpDir, userConfigHash, getExpandedEnvs, runtimeDefineVars = {} } = options;
   const {
     mode,
-    define = {},
     externals = {},
     publicPath = '/',
     outputDir,
@@ -103,30 +167,9 @@ const getWebpackConfig: GetWebpackConfig = ({ rootDir, config, webpack, runtimeT
   const absoluteOutputDir = path.isAbsolute(outputDir) ? outputDir : path.join(rootDir, outputDir);
   const dev = mode !== 'production';
   const hashKey = hash === true ? 'hash:8' : (hash || '');
-  // formate alias
-  const aliasWithRoot = {};
-  Object.keys(alias).forEach((key) => {
-    const aliasValue = alias[key];
-    aliasWithRoot[key] = (aliasValue && typeof aliasValue === 'string' && aliasValue.startsWith('.')) ? path.join(rootDir, aliasValue) : aliasValue;
-  });
 
-  // auto stringify define value
-  const defineVars = {};
-  Object.keys(define).forEach((key) => {
-    defineVars[key] = JSON.stringify(define[key]);
-  });
-
-  const runtimeDefineVars = {};
-  const RUNTIME_PREFIX = /^ICE_/i;
-  Object.keys(process.env).filter((key) => {
-    return RUNTIME_PREFIX.test(key) || ['NODE_ENV'].includes(key);
-  }).forEach((key) => {
-    runtimeDefineVars[`process.env.${key}`] =
-      /^ICE_CORE_/i.test(key)
-        // ICE_CORE_* will be updated dynamically, so we need to make it effectively
-        ? webpack.DefinePlugin.runtimeValue(() => JSON.stringify(process.env[key]), true)
-        : JSON.stringify(process.env[key]);
-  });
+  const aliasWithRoot = getAliasWithRoot(rootDir, alias);
+  const defineVars = getDefineVars(config, runtimeDefineVars, getExpandedEnvs, webpack);
 
   const lazyCompilationConfig = dev && experimental?.lazyCompilation ? {
     lazyCompilation: {
@@ -149,6 +192,7 @@ const getWebpackConfig: GetWebpackConfig = ({ rootDir, config, webpack, runtimeT
       // https://github.com/vercel/next.js/issues/7178#issuecomment-493048965
       comparisons: false,
       inline: 2,
+      passes: 4,
     },
     mangle: {
       safari10: true,
@@ -159,6 +203,7 @@ const getWebpackConfig: GetWebpackConfig = ({ rootDir, config, webpack, runtimeT
       // Fixes usage of Emoji and certain Regex
       ascii_only: true,
     },
+    module: true,
   }, minimizerOptions);
   const compilation = compilationPlugin({
     cacheDir,
@@ -191,6 +236,7 @@ const getWebpackConfig: GetWebpackConfig = ({ rootDir, config, webpack, runtimeT
     },
     context: rootDir,
     module: {
+      unsafeCache: false,
       parser: {
         javascript: {
           importExportsPresence: 'warn',
@@ -230,7 +276,7 @@ const getWebpackConfig: GetWebpackConfig = ({ rootDir, config, webpack, runtimeT
     },
     watchOptions: {
       // add a delay before rebuilding once routes changed
-      // webpack can not found routes component after it is been deleted
+      // webpack can not be found routes component after it has been deleted
       aggregateTimeout: 200,
       ignored: watchIgnoredRegexp,
     },
@@ -276,6 +322,8 @@ const getWebpackConfig: GetWebpackConfig = ({ rootDir, config, webpack, runtimeT
     plugins: [
       ...plugins,
       ...compilerWebpackPlugins,
+      // @ts-ignore
+      new EnvReplacementPlugin(),
       dev && fastRefresh && new ReactRefreshWebpackPlugin({
         exclude: [/node_modules/, /bundles[\\\\/]compiled/],
         // use webpack-dev-server overlay instead
@@ -410,12 +458,9 @@ const getWebpackConfig: GetWebpackConfig = ({ rootDir, config, webpack, runtimeT
     webpack,
     enableRpx2Vw,
   };
-  const finalWebpackConfig = [configCss, configAssets, ...(configureWebpack || [])]
+  return [configCss, configAssets, ...(configureWebpack || [])]
     .reduce((result, next: ModifyWebpackConfig<Configuration, typeof webpack>) => next(result, ctx), webpackConfig);
-    // TODO: Log webpack config with namespace.
-    // consola.debug('[webpack]', finalWebpackConfig);
-  return finalWebpackConfig;
-};
+}
 
 function getDevtoolValue(sourceMap: Config['sourceMap']) {
   if (typeof sourceMap === 'string') {
@@ -427,7 +472,4 @@ function getDevtoolValue(sourceMap: Config['sourceMap']) {
   return 'source-map';
 }
 
-export {
-  getWebpackConfig,
-  getCompilerPlugins,
-};
+export { getCompilerPlugins };