+++ packages/plugin-icestark/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.2
+
+- [feat] support lifecycle of `bootstrap` for qiankun micro app.
+
+## 1.0.1
+
+- [fix] modify basename when render as a child app.
+
 ## 1.0.0
 
 - [feat] plugin for micro-fontends, support both `Framework app` and `Sub app`.