+++ packages/plugin-pha/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 1.1.1
 
+- [feat] support resource_prefetch to preload resource
 - [feat] support enableExpiredManifest
+- [feat] support configure `enableExpiredManifest`
 
 ## 1.1.0