+++ packages/webpack-config/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.0.5
 
 - [fix] isAppEntry shouldn't include app-work.ts
+- [fix] bump version of `@ice/bundles`(0.1.2 -> 0.1.3)
+- [feat] define runtime value of `import.meta.*`
 
 ## 1.0.4